Refactor: rename Interfaces folders to the Abstractions
