+++ Booking/Hotel.Booking.Domain/Abstractions/ITicketsRepository.cs
@@ -1,6 +1,6 @@
 using Hotel.Booking.Domain.Entities;
 
-namespace Hotel.Booking.Domain.Interfaces;
+namespace Hotel.Booking.Domain.Abstractions;
 
 /// <summary>
 /// Интерфейс репозитория с заявками на бронирование
