+++ Identity/Hotel.Identity.Domain/Services/UserService.cs
@@ -2,10 +2,10 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 using FluentValidation;
+using Hotel.Identity.Domain.Abstractions;
 using Hotel.Identity.Domain.DataObjects;
 using Hotel.Identity.Domain.Entities;
 using Hotel.Identity.Domain.Enums;
-using Hotel.Identity.Domain.Interfaces;
 using Hotel.Identity.Domain.Utils;
 using Hotel.Shared.Exceptions;
 using Microsoft.EntityFrameworkCore;
