+++ Managing/Hotel.Managing.Infrastructure/ServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-using Hotel.Managing.Domain.Interfaces;
+using Hotel.Managing.Domain.Abstractions;
 using Hotel.Managing.Infrastructure.Database;
 using Hotel.Managing.Infrastructure.RabbitConsumers;
 using Hotel.Managing.Infrastructure.Services;
